pyproject.toml
[project]
name = "policy-forge"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "typer (>=0.16.0,<0.17.0)",
    "rich (>=14.0.0,<15.0.0)",
    "openai (>=1.84.0,<2.0.0)",
    "pydantic (>=2.11.5,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "markdownify (>=1.1.0,<2.0.0)",
    "python-slugify (>=8.0.4,<9.0.0)"
]

[tool.poetry]
packages = [{include = "policy_forge", from = "src"}]


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
mypy = "^1.16.0"

[tool.poetry.scripts]
policyforge = "policy_forge.__main__:app"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
